requirements.txt
pyrogram>=2.0.106
tgcrypto>=1.2.5
python-dotenv>=1.0.0
asyncio>=3.4.3
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
requests>=2.31.0
motor>=3.3.1
pymongo>=4.6.0 
aiohttp>=3.8.5
dnspython>=2.4.2
